requirements.txt
asgiref==3.3.4
autopep8==1.5.6
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.2.4
django-cities==0.6
django-countries==7.1
django-heroku==0.3.1
django-model-utils==4.1.1
django-phonenumber-field==5.1.0
idna==2.10
phonenumbers==8.12.22
Pillow==8.2.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycodestyle==2.7.0
PyJWT==1.7.1
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
swapper==1.1.2.post1
toml==0.10.2
tqdm==4.60.0
twilio==6.59.0
urllib3==1.26.4
whitenoise==5.2.0
